requirements.txt
annotated-types==0.7.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.2
playwright==1.50.0
playwright-stealth==1.0.6
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
PySocks==1.7.1
pyTelegramBotAPI==4.26.0
pytest==8.3.4
pytest-asyncio==0.25.3
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
selenium==4.28.1
setuptools==75.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
telebot==0.0.5
tiktok-captcha-solver==0.6.2
toml==0.10.2
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.2
wsproto==1.2.0
yt-dlp==2025.1.26
